Restructured util
