+++ web/src/components/Dashboard/Dashboard.tsx
@@ -1,11 +1,14 @@
 import { ChangeEvent, Fragment, useContext, useEffect, useState } from 'react'
-import type { ICourseData, ITerm } from '../../utils/interfaces'
+import type { ICourseData, ITerm } from '../../utils/interfaces/interfaces'
 import CourseCard from './CourseCard/CourseCard.component'
-import { subject_abbr } from '../../utils/subject_abbr.data'
 import UtilCard from './UtilCard/UtilCard.component'
 import { CourseDataContext } from '../../contexts/CourseData.context'
+import { UTIL_CARD_MAP } from '../../utils/data/utilcard.data'
+import { process_course_name, process_search_string } from '../../utils/functions/course_name_util'
+import { TERMS } from '../../utils/data/terms.data'
 
 const Dashboard = () => {
+    // context & state hooks
     const { courses } = useContext(CourseDataContext)
     const [search_string, set_search_string] = useState('')
     const [courses_this_term, set_courses_this_term] = useState<
@@ -15,78 +18,9 @@ const Dashboard = () => {
         Array<ICourseData>
     >([])
 
-    /** Processes search string for searching
-     * abbr => full course name
-     * @example 'cs189' => 'compsci 189'
-     * @returns full course name + number
-     */
-    const process_search_string = (search_string: string): string => {
-        var returned_string = search_string
-            .toLowerCase()
-            .replace(/[^a-z0-9\s,]/g, '')
+    const [selected_term, set_selected_term] = useState<ITerm>(TERMS[0])
 
-        for (const key in subject_abbr) {
-            returned_string = returned_string.replace(
-                new RegExp(`^${key}`),
-                subject_abbr[key]
-            )
-        }
-
-        return standardize_course_name(returned_string)
-    }
-
-    /** Processes course_name for searching
-     *
-     * @example 'DATA C100' => 'data 100'
-     * @returns lowercase course name with leading c in course number removed
-     */
-    const process_course_name = (course_name: string): string => {
-        return standardize_course_name(course_name)
-    }
-
-    /** Standardize course name / search string
-     * removes special number tags like ['c', 'w', 'n']
-     * @returns
-     */
-    const standardize_course_name = (course_name: string): string => {
-        const resplit = course_name
-            .toLowerCase()
-            .replace(' ', '')
-            .replace(/([a-z]*)(.*)/, '$1 $2')
-            .split(' ')
-
-        var dept = resplit[0]
-
-        // NOTE: This is a shortcut for removing trailing c's in department names
-            // If there are more than one c's, they will all be removed to keep consistency
-            // eg. 'DataCCC100' => 'Data100'
-            // It might cause problems in the future if the university change the department names, but it's good for now
-        for (const char of ['c', 'w', 'n']) {
-            const key = new RegExp(`${char}+$`)
-            if (dept.match(key)) {
-                dept = dept.replace(key, '')
-                break
-            }
-        }
-        const num = resplit[1]
-
-        if (dept === '') {
-            return num
-        } else if (num === '') {
-            return dept
-        } else {
-            return dept + ' ' + num
-        }
-    }
-
-    const terms: ITerm[] = [
-        { school_name_and_term: 'UCB Mj01', label: '专业群' },
-        { school_name_and_term: 'UCB Su23', label: 'Summer 2023 课群' },
-        { school_name_and_term: 'UCB Fa23', label: 'Fall 2023 课群' },
-        { school_name_and_term: 'UCB Lf01', label: 'Cal Life' }
-    ]
-
-    const [selected_term, set_selected_term] = useState<ITerm>(terms[0])
+    // effect hooks
 
     // filter courses when selecting new term / category
     useEffect(() => {
@@ -151,7 +85,7 @@ const Dashboard = () => {
                 id="filterBar"
                 className="group grid relative w-fit text-center grid-cols-4 my-8 mx-auto"
             >
-                {terms.map(term => {
+                {TERMS.map(term => {
                     let selected =
                         term.school_name_and_term ===
                         selected_term.school_name_and_term
@@ -189,8 +123,8 @@ const Dashboard = () => {
                 ) : (
                     <UtilCard
                         key={'request'}
-                        label="申请建群"
-                        onClickHandler={() => {}}
+                        label={UTIL_CARD_MAP.add_request.label}
+                        onClickHandler={UTIL_CARD_MAP.add_request.onClickHandler}
                     />
                 )}
