+++ web/src/requests/Course-API/report-missing-class.ts
@@ -1,4 +1,4 @@
-import { MissingRecord } from '../../utils/interfaces';
+import { MissingRecord } from '../../utils/interfaces/interfaces';
 import { basePostRequest } from '../base-requests';
 
 export const reportMissingClass = (
