+++ web/src/utils/data/terms.data.ts
@@ -1,0 +1,8 @@
+import { ITerm } from "../interfaces/interfaces";
+
+export const TERMS: ITerm[] = [
+    { school_name_and_term: 'UCB Mj01', label: '专业群' },
+    { school_name_and_term: 'UCB Su23', label: 'Summer 2023 课群' },
+    { school_name_and_term: 'UCB Fa23', label: 'Fall 2023 课群' },
+    { school_name_and_term: 'UCB Lf01', label: 'Cal Life' }
+]
